// src/app/core/models/mortgage.model.ts
export class Mortgage {
    
    amount: number;
    fakeMortId: string;
    mortgageId: number;
    nepDate: string;
    notiSeen: number;
    qty: string;
    reason: string;
    engDate: string;

    //   mortgageDetailsCollection: MortgageDetailsCollection[];
    //   interestCollection: InterestCollection[];
    // interestReceivedCollection
}